+++ src/main/java/kitchenpos/order/domain/TableGroupSavedEvent.java
@@ -1,4 +1,4 @@
-package kitchenpos.tablegroup.domain;
+package kitchenpos.order.domain;
 
 import java.util.List;
