+++ src/main/java/kitchenpos/order/repository/OrderTableRepository.java
@@ -1,10 +1,10 @@
-package kitchenpos.ordertable.repository;
+package kitchenpos.order.repository;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.order.domain.OrderTable;
 
 public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
     List<OrderTable> findAllByIdIn(List<Long> orderTableIds);
