+++ src/main/java/kitchenpos/order/repository/TableGroupRepository.java
@@ -1,8 +1,8 @@
-package kitchenpos.tablegroup.repository;
+package kitchenpos.order.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kitchenpos.tablegroup.domain.TableGroup;
+import kitchenpos.order.domain.TableGroup;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
 }
