+++ src/test/java/kitchenpos/order/domain/OrderTest.java
@@ -11,7 +11,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import kitchenpos.common.exception.BadRequestException;
-import kitchenpos.ordertable.domain.OrderTable;
 
 class OrderTest {
